feat(media): add default quick trim mode with draggable range selection

+++ app/src/main/java/org/zenconverter/app/conversion/ConversionTaskStore.kt
@@ -67,8 +67,8 @@ data class AudioExportOptions(
 )
 
 data class MediaTrimRange(
-    val startSeconds: Long? = null,
-    val endSeconds: Long? = null
+    val startSeconds: Double? = null,
+    val endSeconds: Double? = null
 ) {
     val isEnabled: Boolean
         get() = startSeconds != null || endSeconds != null
